Pause Popup

+++ Test2/core/src/com/test2/game/library/Button.java
@@ -1,8 +1,8 @@
 package com.test2.game.library;
 
 public class Button {
-    int x=0,y=0,width=0,height=0;
-   Button(int _x,int _y, int _width, int _height) {
+    float x=0,y=0,width=0,height=0;
+   Button(float _x,float _y, float _width, float _height) {
        x=_x;
        y=_y;
        width=_width;
